LayoutTests/fast/events/pageshow-pagehide-on-back-cached-with-frames-expected.txt

CONSOLE MESSAGE: line 21: ***Top level frame being parsed for the initial page load***
CONSOLE MESSAGE: line 21: Subsubframe window.onload
CONSOLE MESSAGE: line 21: Subsubframe window.onpageshow, target = [object HTMLDocument], persisted = false
CONSOLE MESSAGE: line 21: Subframe window.onload
CONSOLE MESSAGE: line 21: Subframe window.onpageshow, target = [object HTMLDocument], persisted = false
CONSOLE MESSAGE: line 21: Subsubframe2 window.onload
CONSOLE MESSAGE: line 21: Subsubframe2 window.onpageshow, target = [object HTMLDocument], persisted = false
CONSOLE MESSAGE: line 21: Subframe2 window.onload
CONSOLE MESSAGE: line 21: Subframe2 window.onpageshow, target = [object HTMLDocument], persisted = false
CONSOLE MESSAGE: line 21: Main frame window.onload
CONSOLE MESSAGE: line 21: Main frame window.onpageshow, target = [object HTMLDocument], persisted = false
CONSOLE MESSAGE: line 21: ***Navigating top-level frame to a page that will immediately navigate back to this one***
CONSOLE MESSAGE: line 21: Main frame window.onpagehide, target = [object HTMLDocument], persisted = true
CONSOLE MESSAGE: line 21: Subframe window.onpagehide, target = [object HTMLDocument], persisted = true
CONSOLE MESSAGE: line 21: Subsubframe window.onpagehide, target = [object HTMLDocument], persisted = true
CONSOLE MESSAGE: line 21: Subframe2 window.onpagehide, target = [object HTMLDocument], persisted = true
CONSOLE MESSAGE: line 21: Subsubframe2 window.onpagehide, target = [object HTMLDocument], persisted = true
CONSOLE MESSAGE: line 21: Subsubframe window.onpageshow, target = [object HTMLDocument], persisted = true
CONSOLE MESSAGE: line 21: Subframe window.onpageshow, target = [object HTMLDocument], persisted = true
CONSOLE MESSAGE: line 21: Subsubframe2 window.onpageshow, target = [object HTMLDocument], persisted = true
CONSOLE MESSAGE: line 21: Subframe2 window.onpageshow, target = [object HTMLDocument], persisted = true
CONSOLE MESSAGE: line 21: Main frame window.onpageshow, target = [object HTMLDocument], persisted = true
Test pageshow/pagehide event behavior when navigating away from a page with frames, putting the page in the page cache, then back to it.
 
